Project/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid==2.2.5
click==8.1.2
ecdsa==0.13.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
future==0.17.1
isort==4.3.18
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.4.0
MarkupSafe==2.1.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.16.0
typed-ast==1.4.2
Werkzeug==2.1.1
wrapt==1.11.1
Flask-Cors==3.0.8